move refactored types
