+++ midix/src/parser/reader/error.rs
@@ -21,18 +21,18 @@ pub(crate) enum ParserError {
 
 pub type ParseResult<T> = Result<T, ParserError>;
 
-pub(super) fn unexp_eof() -> io::Error {
+pub(crate) fn unexp_eof() -> io::Error {
     io::Error::new(ErrorKind::UnexpectedEof, "Read past the end of the file")
 }
 
-pub(super) fn inv_data<R>(reader: &mut Reader<R>, v: impl fmt::Display) -> io::Error {
+pub(crate) fn inv_data<R>(reader: &mut Reader<R>, v: impl fmt::Display) -> io::Error {
     reader.set_last_error_offset(reader.buffer_position());
     io::Error::new(
         ErrorKind::InvalidData,
         format!("Cursor at {}: {}", reader.buffer_position(), v),
     )
 }
-pub(super) fn inv_input<R>(reader: &mut Reader<R>, v: impl fmt::Display) -> io::Error {
+pub(crate) fn inv_input<R>(reader: &mut Reader<R>, v: impl fmt::Display) -> io::Error {
     reader.set_last_error_offset(reader.buffer_position());
     io::Error::new(
         ErrorKind::InvalidInput,
